+++ endpoints/sms-gateway.php
@@ -21,7 +21,7 @@ if (str_exists(strtoupper($address), strtoupper($sms_helpline_keyword))) {
     } else {
         ?>
         <Sms><?php echo word('please_send_a_message_formatting_as') ?> "<?php echo $sms_helpline_keyword?>", <?php echo word('followed_by_your_location')?>, <?php echo word('for') ?> <?php echo word('someone_to_talk_to')?>.</Sms>
-    <?php   }
+    <?php }
 } else if (str_exists(strtoupper($address), strtoupper('jft')) && json_decode(setting('jft_option'))) {
     $jft_chunks = get_jft(true);
     for ($i = 0; $i < count($jft_chunks); $i++) {