+++ RELEASENOTES.md
@@ -4,8 +4,9 @@
 * Introducing Yap installer. [#373]
 * Added custom extensions feature that allows for setting up arbitrary call forwards and/or prompt playbacks. (https://github.com/bmlt-enabled/yap/wiki/Custom-Extensions) [#355]
 * Added an option to combine SMS messages into a single SMS versus individual ones.
+* Added virtual links for those meetings tagged with VM for location-centric virtual meetings.
 * Added time format customization option.
-* Added meeting results voice supression option.
+* Added meeting results voice suppression option.
 * Added the ability to call in and listen to voicemail messages. [#384]
 * Dropping support for Internet Explorer 11. [#389]